HAD ERROR

+++ backend_laravelPHP/database/migrations/2024_04_06_202357_create_employees_table.php
@@ -4,12 +4,15 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+
 return new class extends Migration
 {
     public function up(): void
     {
+       
         // Create the 'employees' table
         Schema::create('employees', function (Blueprint $table) {
+            $table->engine = 'InnoDB'; 
             $table->id();
             $table->string('employee_firstname')->nullable();
             $table->string('employee_middlename')->nullable();
@@ -33,21 +36,25 @@ return new class extends Migration
             $table->string('employee_pagibig_no', 255)->nullable();
             $table->string('employee_philhealth_no', 255)->nullable();
             $table->string('employee_tin_no', 255)->nullable();
+            $table->foreign('access_type_id')
+            ->references('id')->on('access_types')
+            ->onDelete('set null');
             $table->unsignedBigInteger('employee_department_id')->nullable();
             $table->unsignedBigInteger('employee_civil_status_id')->nullable();
-            $table->unsignedBigInteger('access_type_id')->nullable(); 
             $table->rememberToken();
             $table->timestamps();
 
-            // Foreign key definition for 'access_type_id'
-            $table->foreign('access_type_id')
-                ->references('id')
-                ->on('access_types')
-                ->onDelete('set null');
         });
 
         // Add foreign key constraints after table creation
         Schema::table('employees', function (Blueprint $table) {
+            if (Schema::hasTable('access_types')) {
+                $table->foreign('access_type_id')
+                    ->references('id')
+                    ->on('access_types')
+                    ->onDelete('set null');
+            }
+
             if (Schema::hasTable('departments')) {
                 $table->foreign('employee_department_id')
                     ->references('id')
@@ -62,12 +69,14 @@ return new class extends Migration
                     ->onDelete('set null');
             }
         });
+
     }
 
     public function down(): void
     {
         // Drop the foreign key constraints and the 'employees' table
         Schema::table('employees', function (Blueprint $table) {
+            $table->dropForeign(['access_type_id']);
             $table->dropForeign(['employee_department_id']);
             $table->dropForeign(['employee_civil_status_id']);
         });
